Revert "Replace whole site with under construction communicate"
This reverts commit 6a57e1396d8b65ed20c98c92177a0d8232bcb7c5.
