--- src/App.jsx
@@ -1,12 +1,0 @@
-import Logo from "./assets/logo.png";
-
-export default function App() {
-  return (
-    <div className="under-construction">
-      <div className="hero-logo">
-        <img src={Logo} alt="Logo parku dinozaurów Łeba" />
-      </div>
-      <h1>Strona w budowie</h1>
-    </div>
-  );
-}
